Move clean adduct and looks_like_adduct to clean adduct
